+++ cpp-cgdk/model/World.cpp
@@ -58,7 +58,7 @@ const vector<Facility>& World::getFacilities() const {
 }
 
 Player World::getMyPlayer() const {
-    for (int playerIndex = players.size() - 1; playerIndex >= 0; --playerIndex) {
+    for (size_t playerIndex = players.size() - 1; playerIndex >= 0; --playerIndex) {
         Player player = players[playerIndex];
         if (player.isMe()) {
             return player;
@@ -69,7 +69,7 @@ Player World::getMyPlayer() const {
 }
 
 Player World::getOpponentPlayer() const {
-    for (int playerIndex = players.size() - 1; playerIndex >= 0; --playerIndex) {
+    for (size_t playerIndex = players.size() - 1; playerIndex >= 0; --playerIndex) {
         Player player = players[playerIndex];
         if (!player.isMe()) {
             return player;
